refactor: take out the handlers take the handlers out of the router
