+++ src/Router/handlers/index.ts
@@ -1,0 +1,5 @@
+export { getAllUsersDB } from './getAllUsersDB.js';
+export { addUserDB } from './addUserDB.js';
+export { getUserDB } from './getUserDB.js';
+export { putUserDB } from './putUserDB.js';
+export { deleteUserDB } from './deleteUserDB.js';
